requirements.txt
openai==1.33.0
ruamel.yaml==0.18.6
python-dotenv==1.0.1
pyodbc==5.1.0
pandas==2.2.2